wip: analysis/binary/IR/disasm/decompile/script + agent + UI + docs
Cross-cutting in-flight changes:
  - core/analysis: pipeline scope plumbing, indirect-call resolver tweaks
  - core/binary: ELF/PE/PDB/minidump/raw-regions loader hardening
  - core/common/annotations: API surface fills for the agent harness
    promote path
  - core/decompile/emitter: refined value-type for inferred function
    return; struct-field annotation flow
  - core/disasm: ARM decoder coverage extensions; x86 decoder edges
  - core/ir: lifter/passes/ssa cleanup pass tweaks
  - core/script/declarative: parser surface for the new directives
  - cli/info, cli/main: doc/help additions for new flags
  - agent/src/promote: declarative output path
  - ui: renderer + Electron main shims for the plugin platform
  - docs: scripting.md additions
  - tests: declarative + arm64 + annotations_resolver coverage extensions

+++ core/src/binary/pe.cpp
@@ -533,8 +533,13 @@ collect_exports(const Binary& bin,
     // Forwarder exports point *into* the export data directory itself —
     // the RVA there is the "DLL.Symbol" redirect string, not a function
     // address. Skip them instead of emitting a garbage Symbol at a
-    // string VA.
-    const u32 dd_end = dd.virtual_address + dd.size;
+    // string VA. Widen to u64 for the upper-bound check: a hostile PE
+    // with `dd.virtual_address + dd.size` that overflows u32 would
+    // otherwise wrap to a tiny number, causing every export to fail
+    // (or pass) the forwarder test depending on which side the wrap
+    // lands on.
+    const u64 dd_end = static_cast<u64>(dd.virtual_address) +
+                       static_cast<u64>(dd.size);
     const auto is_forwarder = [&](u32 rva) noexcept {
         return rva >= dd.virtual_address && rva < dd_end;
     };
